Prevent cyclic dependencies
JIRA: RAIL-3466
