+++ libs/sdk-ui-dashboard/src/model/index.ts
@@ -12,12 +12,8 @@
  * -  all commands, their types & command factories
  */
 
-export {
-    useDashboardSelector,
-    useDashboardDispatch,
-    DashboardDispatch,
-    DashboardState,
-} from "./state/dashboardStore";
+export { useDashboardSelector, useDashboardDispatch, DashboardDispatch } from "./state/dashboardStore";
+export { DashboardState } from "./state/types";
 
 export { selectDashboardLoading } from "./state/loading/loadingSelectors";
 export { LoadingState } from "./state/loading/loadingState";
