+++ libs/sdk-ui-dashboard/src/model/state/dateFilterConfig/dateFilterConfigSelectors.ts
@@ -1,6 +1,6 @@
 // (C) 2021 GoodData Corporation
 import { createSelector } from "@reduxjs/toolkit";
-import { DashboardState } from "../dashboardStore";
+import { DashboardState } from "../types";
 import invariant from "ts-invariant";
 import { DashboardDateFilterConfigMode, DateFilterGranularity } from "@gooddata/sdk-backend-spi";
 import { IDateFilterOptionsByType } from "@gooddata/sdk-ui-filters";
