+++ libs/sdk-ui-dashboard/src/model/state/filterContext/filterContextSelectors.ts
@@ -1,6 +1,6 @@
 // (C) 2021 GoodData Corporation
 import { createSelector } from "@reduxjs/toolkit";
-import { DashboardState } from "../dashboardStore";
+import { DashboardState } from "../types";
 import invariant from "ts-invariant";
 import {
     FilterContextItem,
